move test cases for RemoteProcessor out
