+++ tests/forte/advanced_pipeline_test.py
@@ -27,11 +27,10 @@ from forte.data.data_pack import DataPack
 from forte.data.multi_pack import MultiPack
 from forte.data.ontology.code_generation_objects import EntryTreeNode
 from forte.data.ontology.top import Generics
-from forte.data.readers import RawDataDeserializeReader, StringReader
+from forte.data.readers import StringReader
 from forte.data.selector import RegexNameMatchSelector
 from forte.pipeline import Pipeline
 from forte.processors.base import PackProcessor
-from forte.processors.misc import RemoteProcessor
 from forte.processors.nlp import ElizaProcessor
 from ft.onto.base_ontology import Utterance
 
@@ -107,7 +106,7 @@ class DummyProcessor(PackProcessor):
 @ddt
 class AdvancedPipelineTest(unittest.TestCase):
     """
-    Test intermediate representation and RemoteProcessor. Here we use eliza
+    Test intermediate representation. Here we use eliza
     pipeline as an example, and all the testcases below are refactored from
     `eliza_test.py`.
     """
@@ -212,56 +211,6 @@ class AdvancedPipelineTest(unittest.TestCase):
         ],
         ["bye", "Goodbye.  Thank you for talking to me."],
     )
-    def test_remote_processor(self, input_output_pair):
-        """
-        Verify RemoteProcessor.
-        """
-        i_str, o_str = input_output_pair
-        service_name: str = "test_service_name"
-        input_format: str = "DataPack"
-
-        # Build service pipeline
-        serve_pl: Pipeline[DataPack] = Pipeline[DataPack]()
-        serve_pl.set_reader(RawDataDeserializeReader())
-        serve_pl.add(DummyProcessor(expected_records=TEST_RECORDS_1))
-        serve_pl.add(UserSimulator(), config={"user_input": i_str})
-        serve_pl.add(DummyProcessor(output_records=TEST_RECORDS_2))
-        serve_pl.add(ElizaProcessor())
-        serve_pl.initialize()
-
-        # Configure RemoteProcessor into test mode
-        remote_processor: RemoteProcessor = RemoteProcessor()
-        remote_processor.set_test_mode(
-            serve_pl._remote_service_app(
-                service_name=service_name, input_format=input_format
-            )
-        )
-
-        # Build test pipeline
-        test_pl: Pipeline[DataPack] = Pipeline[DataPack](
-            do_init_type_check=True
-        )
-        test_pl.set_reader(StringReader())
-        test_pl.add(DummyProcessor(output_records=TEST_RECORDS_1))
-        test_pl.add(
-            remote_processor,
-            config={
-                "validation": {
-                    "do_init_type_check": True,
-                    "input_format": input_format,
-                    "expected_name": service_name,
-                }
-            },
-        )
-        test_pl.add(DummyProcessor(expected_records=TEST_RECORDS_2))
-        test_pl.initialize()
-
-        # Verify output
-        res: DataPack = test_pl.process("")
-        utterance = get_last_utterance(res, "ai")
-        self.assertEqual(len([_ for _ in res.get(Utterance)]), 2)
-        self.assertEqual(utterance.text, o_str)
-
     def _assertEntryTreeEqual(self, root1: EntryTreeNode, root2: EntryTreeNode):
         """
         Test if two `EntryTreeNode` objects are recursively equivalent
